Remove custom CORS handling
