+++ infra/variables.tf
@@ -14,11 +14,6 @@ variable "domain_name" {
   default     = "notes.thalman.org"
 }
 
-variable "allowed_origin" {
-  description = "Allowed origin for CORS headers"
-  type        = string
-  default     = null
-}
 
 variable "domain_name_root" {
   description = "Route53 hosted zone domain"
