+++ packages/backend/src/workspaces.ts
@@ -5,7 +5,7 @@ import { DynamoDB } from 'aws-sdk';
 import { Workspace } from '@sticky-notes/shared';
 import { getUserId, hasWorkspaceAccess } from './auth';
 import { broadcastWorkspaceEvent } from './websocket';
-import { CORS_HEADERS, withErrorHandling } from './cors';
+import { withErrorHandling } from './error';
 
 const TABLE_NAME = process.env.TABLE_NAME as string;
 const db = new DynamoDB.DocumentClient();
@@ -52,7 +52,6 @@ export const createWorkspace = withErrorHandling(async (event) => {
 
   return {
     statusCode: 201,
-    headers: CORS_HEADERS,
     body: JSON.stringify(workspace),
   };
 });
@@ -65,7 +64,7 @@ export const getWorkspace = withErrorHandling(async (event) => {
   const userId = getUserId(event);
   const id = event.pathParameters?.id;
   if (!id) {
-    return { statusCode: 400, headers: CORS_HEADERS, body: 'Missing id' };
+    return { statusCode: 400, body: 'Missing id' };
   }
 
   const data = await db
@@ -74,16 +73,15 @@ export const getWorkspace = withErrorHandling(async (event) => {
   const workspace = data.Item as Workspace | undefined;
 
   if (!workspace) {
-    return { statusCode: 404, headers: CORS_HEADERS, body: 'Workspace not found' };
+    return { statusCode: 404, body: 'Workspace not found' };
   }
 
   if (!hasWorkspaceAccess(workspace, userId)) {
-    return { statusCode: 403, headers: CORS_HEADERS, body: 'Forbidden' };
+    return { statusCode: 403, body: 'Forbidden' };
   }
 
   return {
     statusCode: 200,
-    headers: CORS_HEADERS,
     body: JSON.stringify(workspace),
   };
 });
@@ -96,7 +94,7 @@ export const updateWorkspace = withErrorHandling(async (event) => {
   const userId = getUserId(event);
   const id = event.pathParameters?.id;
   if (!id) {
-    return { statusCode: 400, headers: CORS_HEADERS, body: 'Missing id' };
+    return { statusCode: 400, body: 'Missing id' };
   }
 
   const existing = await db
@@ -105,11 +103,11 @@ export const updateWorkspace = withErrorHandling(async (event) => {
   const workspace = existing.Item as Workspace | undefined;
 
   if (!workspace) {
-    return { statusCode: 404, headers: CORS_HEADERS, body: 'Workspace not found' };
+    return { statusCode: 404, body: 'Workspace not found' };
   }
 
   if (!hasWorkspaceAccess(workspace, userId)) {
-    return { statusCode: 403, headers: CORS_HEADERS, body: 'Forbidden' };
+    return { statusCode: 403, body: 'Forbidden' };
   }
 
   const updates: Partial<Workspace> = event.body ? JSON.parse(event.body) : {};
@@ -151,7 +149,6 @@ export const updateWorkspace = withErrorHandling(async (event) => {
 
   return {
     statusCode: 200,
-    headers: CORS_HEADERS,
     body: JSON.stringify(updated.Attributes),
   };
 });
@@ -163,7 +160,7 @@ export const deleteWorkspace = withErrorHandling(async (event) => {
   const userId = getUserId(event);
   const id = event.pathParameters?.id;
   if (!id) {
-    return { statusCode: 400, headers: CORS_HEADERS, body: 'Missing id' };
+    return { statusCode: 400, body: 'Missing id' };
   }
 
   const res = await db
@@ -172,11 +169,11 @@ export const deleteWorkspace = withErrorHandling(async (event) => {
   const workspace = res.Item as Workspace | undefined;
 
   if (!workspace) {
-    return { statusCode: 404, headers: CORS_HEADERS, body: 'Workspace not found' };
+    return { statusCode: 404, body: 'Workspace not found' };
   }
 
   if (!hasWorkspaceAccess(workspace, userId)) {
-    return { statusCode: 403, headers: CORS_HEADERS, body: 'Forbidden' };
+    return { statusCode: 403, body: 'Forbidden' };
   }
 
   await db
@@ -185,7 +182,6 @@ export const deleteWorkspace = withErrorHandling(async (event) => {
 
   return {
     statusCode: 204,
-    headers: CORS_HEADERS,
     body: '',
   };
 });
@@ -205,7 +201,7 @@ export const listWorkspaces = withErrorHandling(async (event) => {
   const ids = Array.from(new Set([...owned, ...contrib]));
 
   if (ids.length === 0) {
-    return { statusCode: 200, headers: CORS_HEADERS, body: '[]' };
+    return { statusCode: 200, body: '[]' };
   }
 
   const res = await db
@@ -221,7 +217,6 @@ export const listWorkspaces = withErrorHandling(async (event) => {
 
   return {
     statusCode: 200,
-    headers: CORS_HEADERS,
     body: JSON.stringify(workspaces),
   };
 });
